+++ frontend/src/core/ai/context/context.ts
@@ -1,11 +1,10 @@
 /* Copyright 2024 Marimo. All rights reserved. */
+
 import { allTablesAtom } from "@/core/datasets/data-source-connections";
 import type { JotaiStore } from "@/core/state/jotai";
 import { variablesAtom } from "@/core/variables/state";
-import {
-  TableContextProvider,
-  VariableContextProvider,
-} from "./providers/context-providers";
+import { TableContextProvider } from "./providers/tables";
+import { VariableContextProvider } from "./providers/variable";
 import { AIContextRegistry } from "./registry";
 
 export function getAIContextRegistry(store: JotaiStore) {
