+++ frontend/src/core/ai/context/providers/common.ts
@@ -1,0 +1,7 @@
+/* Copyright 2024 Marimo. All rights reserved. */
+
+export const Boosts = {
+  LOCAL_TABLE: 5,
+  REMOTE_TABLE: 4,
+  VARIABLE: 3,
+} as const;
